fixed casting bug in test

+++ src/test/java/org/mthree/service/ForecastServiceImplTest.java
@@ -37,7 +37,7 @@ class ForecastServiceImplTest {
         MockitoAnnotations.openMocks(this);
         sampleTransaction = new Transaction();
         sampleTransaction.setPlaidTransactionId("plaidTransactionId");
-        sampleTransaction.setId(1);
+        sampleTransaction.setId((long) 1);
         sampleTransaction.setCategory("Food");
         sampleTransaction.setType("Place");
         sampleTransaction.setName("KFC");
